server/requirements.txt
inputs==0.5
pycapnp==2.0.0
